pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "adversarial-wiki"
version = "0.1.0"
description = "Two opposing knowledge bases that debate each other, surfacing hidden assumptions"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "trafilatura>=1.12.0",
    "tavily-python>=0.5.0",
    "rank-bm25>=0.2.2",
]

[project.scripts]
adversarial-wiki = "adversarial_wiki.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/adversarial_wiki"]
